fix: split exports for context-editor
