+++ packages/context-editor/package.json
@@ -63,24 +63,9 @@
 	"preconstruct": {
 		"entrypoints": [
 			"index.ts",
-			"schemas/index.ts"
+			"schemas/index.ts",
+			"utils/index.ts"
 		]
-	},
-	"exports": {
-		".": {
-			"module": "./dist/context-editor.esm.js",
-			"default": "./dist/context-editor.cjs.js"
-		},
-		"./schemas": {
-			"module": "./schemas/dist/context-editor-schemas.esm.js",
-			"default": "./schemas/dist/context-editor-schemas.cjs.js"
-		},
-		"./package.json": "./package.json",
-		"./style.css": {
-			"module": "./src/style.css",
-			"default": "./src/style.css"
-		},
-		"./tailwind.config.js": "./tailwind.config.js"
 	},
 	"prettier": "@pubpub/prettier-config"
 }
