+++ packages/context-editor/src/index.ts
@@ -2,6 +2,3 @@ import ContextEditor from "./ContextEditor";
 
 export * from "./ContextEditor";
 export { ContextEditor };
-export { baseSchema } from "./schemas";
-export { docHasChanged } from "./utils/hasChanged";
-export { getPubValues } from "./utils/pubValues";
