+++ packages/context-editor/utils/package.json
@@ -1,0 +1,4 @@
+{
+	"main": "dist/context-editor-utils.cjs.js",
+	"module": "dist/context-editor-utils.esm.js"
+}
